Removing unnecessary files
Most of the files that come with the default project template that Jets
creates with the `new` command aren't needed for this project (views,
templates, routes, JS/CSS/HTML config...). Therefore, we'll delete it

We have also removed the libraries that we won't be using (e.g. the ones
related with DynamoDB, or with JS like webpack), and installed the PG
gem to allow using PostgreSQL as our database