Migrating Meal and Injection models to Dynamoid
The models preserve the same fields as before, but we have converted
injection_type and meal_type to strings, as we can't use enums anymore.

We have added validations in the models to check the presence of the
fields that are required both in the Injection (injection_type and
units) and in the Meal (food and meal_type) models.

Also, the DB tables have been renamed (because Dynamoid uses a
different convention than ActiveRecord), but it's only an internal
change so there is no trace of it in the code