Creating Injection-related logic
We have followed the Hexagonal architecture pattern by dividing the
responsibilities into the following components:

- InjectionParser: that given an SNS message containing information of
an injection, returns a hash with the attributes units, injection_type
and notes. This class inherits from the newly created SnsParser, that
gets the message (JSON encoded as string), parses it and converts it
into a hash.
- CreateInjectionJob: this job, given an event, creates an injection
with the correspondent data. Part of the creation process involves the
use of InjectionParser.

Also, as we did with the models, we've namespaced the classes into the
Health module