Publishing injection data on CloudWatch
Creating a class to act as a wrapper to CloudWatch, and adding a method
to specifically publish an injection. This method contains all the logic
about what information should be published, which Namespace we'll be
using, which dimensions and units, etc.

In the future we may want to have this logic encapsulated independently,
but for the moment it's fine