Publishing metrics after creating the injection
Adding extra logic in CreateInjectionJob to publish the injection metric
after we create it.

We have also done some re-ordering in the lines to make it clearer how
it works: the sns_event triggers the create injection method (as it's
the first one that goes after it). We have removed the blank line
between them to make this more obvious