Ensuring that the metric value is treated as a float
In some tests we discovered that, when trying to publish the metric
value, it was considered a BigInt instead of a float. This caused errors
because CloudWatch service doesn't know how to handle BigInts. In order
to prevent this, we'll manually convert the metric to a float.

We don't expect to publish expenses outside of the float value, so
converting the BigInt to Float shouldn't be a problem at all