Creating Airtable::Expense model
Creating a model to represent the expenses that we'll be tracking in
Airtable. As we did with `Ynab::TransactionData`, we have added a
`.from_expense` method to create the expenses easily from a
`Finance::Expense` object.

As `Airtable::Expense` is an Airrecord model, the way of reading and
setting the attributes it's hash-like, i.e. `object[field] = value`. In
order to make the operations with the object easier, we've added helper
methods to read and write the attributes with dot notation
(`object.field = value`)