Adding Category mapping
Creating a module `Finance::ExpenseCategories` to store the possible
categories that a expense can belong to.

Creating a module `Airtable::Categories` in order to contain the mapping
logic. It has a `MAPPING` field with the equivalencies between the
ExpenseCategories and Airtable's categories

Additionally, we also modified the YNAB implementation to follow the
same strategy. We verified that the tests still pass, which means that
the change was correct