Defining Airtable::Expense attributes with metaprogramming
Instead of having to define the getters and setters for each of the
attributes in the Airtable::Expense class, we have added to the class a
hash that maps attributes to the corresponding column names. Then we use
metaprogramming, more specifically the method `define_method`, to create
the getter and setters automatically for all the keys in the previous
map.

The verification that this feature works is the fact that we didn't make
any change to the unit tests, and they are still passing