Saving Expense in Airtable + Minor fixes
Saving the expense in Airtable when we create it from a Finance::Expense

Additionally, setting the base and the table correctly in the model.
This problem comes from a misunderstanding from my side: what I thought
that was the Table is actually the Base (what would be equivalent to an
Excel File) and the Table is actually the Sheet inside that Base.

Finally, creating an initializer to configure the Airtable client when
our app starts