Creating a model to represent ExpenseMetrics
Instead of directly publishing the metric information as a Hash, we've
created classes that inherit from `Aws::CloudWatch::Types::MetricDatum`
(the class used by CloudWatch for representing metrics). This will allow
us to add specific logic on top of it, and will also simplify the
application logic (e.g. for comparing the metrics that get published or
for testing)