Mapping the expense categories + Tests
Adding a mapping to Expense Categories, to allow us to decouple the
category in the Expense with the Dimension in the CloudWatch Metric

Additionally, we have changed the Dimensions. While this will cause us
not to be able to see the same data as before, it will also be
beneficial as it uses more readable values for them.