Validating the expense category + Case-insensitive categories
Two related changes here:

1) **Validating categories**: as there is a limited amount of valid
expense categories, we have added a validation to the `Finance::Expense`
model to be sure that only valid expenses are registered. If an invalid
category is set, trying to save the expense will raise a
`Dynamoid::Errors::DocumentNotValid` error

2) **Converting the category to lowercase**: In order to avoid getting
validation errors only because the user entered the category in a case
that isn't the default one (e.g. "Supermercado" instead of
"supermercado"), we have overriden the default setter to convert
the value to lowercase before assigning it

Both features have its corresponding tests added in
`CreateExpenseJob` specs. We have also made a minor modification in YNAB
TransactionData spec, as it does not makes sense now to have a category
that does not exist (it will raise an error before we get there).
Instead, we'll consider it as not mapped to a YNAB category