Adding TelegramBot class
Adding a class to send messages to a specific Telegram chat. The
functionality is quite simple: we implemented a method to send a
message to the corresponding chat. We have set the `bot` field as
private in order to allow accessing it only from the specific methods,
hiding undesired side-effects