Creating BankTransaction model
Creating a model to represent a bank transaction. All the banks that we
integrate with the system will use the same model.

The `internal_id` field will be used to uniquely represent a transaction
within a specific bank. It will allow us to distinguish, when we
retrieve the bank information again, if the transaction is already in
the system or not