Creating a service to interact with Openbank API
Finance::Openbank::Service takes care of interacting with Openbank's API
to retrieve the transactions on a specific date range. It also handles
the authentication.

We have implemented specific logic to authenticate and retry if the
request returned a 401 Unauthorized code, and also to retry up to 3
times if we timeout during the request. It seems that there are some
ocassions where Openbank's API takes a lot to answer and the request
timeouts, but retrying afterwards succeeds almost instantly (so it's not
a matter of increasing the default timer).