Adding CloudWatch Alarms to monitor correct metrics publication
Creating a CloudFormation stack with CW Alarms that we'll use to monitor
the correct publication of metrics in our app.

We'll be verifying that, for every datapoint that we publish periodically
(the two types of insulin and the 4 types of expense), there's at least
one datapoint per hour. For insulin we'll have a different alarm for
each type, while for expenses we'll have a single alarm for all of them

For the moment, the alarms are not associated to any action. We'll add
that later