Notifying of Openbank transactions
Adding a set of jobs to retrieve information about the recent bank
transactions on Openbank and sending their information via Telegram.

We have separated the logic into two different jobs:
- **RetrieveYesterdayTransactionsJob**: This job will connect with
Openbank and obtain the information about the bank transactions that
happened the previous day and will store them in the system.
- **NotifyYesterdayTransactionsJob**: This job will query yesterday's
`Finance::BankTransactions` from DynamoDB and send a Telegram message
with its information

The idea of having this as two separate jobs is to decouple the
obtention of information and its notification.

Additionally, while we currently only obtain the information from
Openbank, we are planning on extending this in the future with more Bank
integrations