Error fixing in RetrieveYesterdayTransactionsJob
Two changes:
1) Using `.blank?` instead of `.empty?` on `transactions_data`. This
will prevent errors if it's `nil` (see #25). While it doesn't address
the root cause, at least we mitigate the issue.
2) Modifying the date we retrieve transactions for to make sure that we use it on
UTC. As the Lambda will run on a time close to midnight, we want to be
sure that we don't have problems caused by timezones, so we'll get
yesterday's date after getting the current time in UTC