Disabling Markdown mode in Telegram
In some of the messages we send, part of the text comes from external
sources (e.g. the name of a bank transaction) and we can't control them.
If it contains characters used in Markdown (like * or _) it will cause
Telegram to fail when publishing the message. In order to avoid it,
we'll just disable Markdown mode for the moment