Using Adapters to convert objects to Metrics
Changing the approach from having specific models for metrics to having
an adapter that implements the interface expected by PutMetricData.
These adapters will implement the fields used by CW MetricDatum +
Namespace.

Additionally, we are adding the metric adapter in the model and
delegating the corresponding methods to it, to make the interface more
intuitive