Adding concurrenct processing to the Dexcom job
As we can have a lot of records to process and they are all independent
from each other, we'll process them concurrently.

We do this with the `parallel` gem, and have allocated 4 threads to the
job. In order to make sure that we don't encounter bottlenecks in
DynamoDB, we have also increased the write capacity of the table to 4.

Finally and to be safe, we'll rescue the job from the exception that
Dynamo raises when we are out of capacity and retry that specific
glucose value