Integrating with Fintonic to retrieve Bankia movements
Creating a class to connect with Fintonic's API to retrieve movements
from my Bankia bank account. The class structure is very similar to the
one used to integrate with Openbank.

The interface exposed doesn't make any mention to the fact that we are
using Fintonic underneath. This has been done on purpose, so in case we
ever connect directly with Bankia we don't need to change any external
class