Refactoring transaction retrieval logic
Modifying Finance::RetrieveYesterdayTransactionsJob to avoid repeating
the logic for each new bank added. Instead, we'll take advantage of the
fact that we can infer the classes from the bank name, so we'll
initialize the objects we need by composing the class name directly.

This will let us extend the bank logic to new banks easily, just by
adding the name of the bank to the `BANKS` array