Downgrading puma and byebug to use versions available in Lambdagems
puma and byebug require native extensions to work, which means that they
have to be compiled on an environment similar to where they will be run.
Jets manages this by using www.lambdagems.com, a service that contains
gems already compiled for Lambda.

However, the last versions of puma and byebug are not yet available in
Lambdagems, and therefore the deploy fails. We'll pin them to the
previous version that we were using until they are available.