Add new index depending on transaction date
Creating two new fields, `year_month` and `day`, to be able to group
better the BankTransaction objects. We'll use `year_month` as the hash
key for the index and `day` as the range key. As we have approximately the same number of transactions per month, this
will prevent hot partition problems

Also, we have modified NotifyYesterdayTransactionsJob to execute the
query using the index