Forcing json gem to >= 2.3.1
Due to a security fix, json 2.3.0 is not available in Lambda, and having
it on a Gemfile will cause the app not to run. Therefore, we'll set it
to be at least 2.3.1 to be sure that we don't encounter this problem
during deployments.

More information: https://github.com/tongueroo/jets/pull/470