AIB fixes - Modifying alarm + Adjusting 'to' date
Modifying alarm: The metric tracked by this alarm is sparse, meaning
that it only emits datapoints when there's something wrong. If we
configure it to treat missing data as Not Breaching, any 5 minute period
without datapoints will make the alarm transition to OK, when what we
actually want is to keep it on the same state until a correct datapoint
is published. To make the alarm transition to the OK state, we'll always
emit themetric, and it will be the value what will indicate if there have
been errors or not. Still it will be sparse as this logic is expected to
execute only once per day.

Adjusting 'to' date: TrueLayer considers the end of the range to not to be
included, so to get its movements we need to ask for the transactions up
to the following day