Dividing Alarms in different stacks
Creating different CloudFormation stacks for the CloudWatch Alarms depending on its goal. For the moment we'll have two different stacks: `Alarms-Infrastructure` (the existing stack renamed, to monitor the service health) and `Alarms-Health` (to monitor health-related metrics, like glucose levels or amount of insulin)

The second stack includes a new alarm, `NoBasalInsulinAlarm`, that
tracks if I've taken any basal insulin in the last 24h. It can be
helpful to make sure that I don't miss taking the insulin or tracking
the data. While the evaluation period is not optimal (it will fire if I
one day I take the insulin later than the previous one), CW Alarms don't
allow evaluation periods greater than a day ([link to the docs, see
"Period"](https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_PutMetricAlarm.html))
so this is the best we can do

Finally, we've also modified the script to deploy to prod to make it
more generic so that it deploys all the CloudFormation stacks