Changing the alert message when an alarm transitions to OK
Changing AlarmReceivedJob to use a different message header when the
alarm transitions to OK state. This will allow users to easily
differentiate between messages where the alarm fires and when it gets
back to normal, specially when only the start of the message can be seen
(e.g. on a push notification)