MVP: Job that executes when a new Bank Transaction is created
Creating a job that will execute each time that a job transaction is
created. For the moment this job will only send a Telegram message with
information about the bank transaction. We want to use this as a proof
of concept of how DynamoDB events work, and if they would be suitable
for the use case we are planning on #40