Updating Jets + Deleting unused code
Updating Jets to the latest version, 2.3.17. With it, updating also its
dependencies.

Deleting unused code related with Dexcom API. This code was part of a
PoC that we wanted to use to integrate with Dexcom API via OAuth, but
due to several restrictions (mostly related with complexity/lack of
documentation of the API + the fact that it only offers access to USA
accounts) it wasn't completed. This code needed to be refactored due to
breaking changes on how Jets > 2.0.0 handles routing, and as the code is
unused there is no best refactor than refactor by fire.