Sending a Telegram message for each bank transaction that is created
On 995d05f we added a job that triggers when new BankTransactions are
created. However, we set it to only send a message per each trigger,
when a single trigger can contain several items if all of them were
created closely in time.

This commit modifies the BankTransactionCreatedJob to send a Telegram
message per each bank transaction instead