Adding Postgres to CircleCI + Creating docker-compose for local testing
Adding a postgres image to the CircleCI environments to use it when we
include SQL in the codebase. Additionally, to simplify the local setup,
created a docker-compose file with the DB dependencies. Once it's
started we'll be able to run the tests without having to worry about any
other configuration in the host machine