Migrating Health::Meal model to SQL
Modifying the Health::Meal model to be an ActiveRecord model instead of
a Dynamoid one. Creating the DB migration and the DB schema with the new
modelling.

The new model will have the same fields as the old one, and therefore we
don't need to change any additional part of the codebase