Migrating Meals model to SQL (#52)
* Preparing project to use SQL DBs

Adding dependencies and configuration to use SQL databases in the
project

1) Added a config/database.yml file with the DB configuration
(currently, just for Dev and Test environments)
2) Added pg gem to enable Postgres integration
3) Creating a Rake task to create DB and DB tables before starting the
tests
4) Adding the annotate gem + Rake task to annotate the model files with
the schema configuration

* Adding Postgres to CircleCI + Creating docker-compose for local testing

Adding a postgres image to the CircleCI environments to use it when we
include SQL in the codebase. Additionally, to simplify the local setup,
created a docker-compose file with the DB dependencies. Once it's
started we'll be able to run the tests without having to worry about any
other configuration in the host machine

* Running Postgres on port 5432 to allow CircleCI integration

CircleCI doesn't allow to modify port mapping on Docker images for your
build. As the Postgres port is not configurable via environment
variables or commands, the only option we have left is to run it on the
port used by default by the image, 5432

* Migrating Health::Meal model to SQL

Modifying the Health::Meal model to be an ActiveRecord model instead of
a Dynamoid one. Creating the DB migration and the DB schema with the new
modelling.

The new model will have the same fields as the old one, and therefore we
don't need to change any additional part of the codebase

* Adding configuration for production DB + Running migrations on deploy
script