Automating BankTransaction classification
Extending the ExpenseClassification Rules to add an `expense_category`
method that returns the corresponding ExpenseCategory for that expense.

Creating the Finance::CategoryClassifier class, that given a specific
BankTransaction, analyses all the rules and returns the matching
category. To do this, we have created a `BankTransactionExpenseAdapter`
class, that allows us to access the BankTransaction attributes using the
Expense interface.