Implementing Automated Transaction Classification (#56)
## Description
Related with #40

This PR implements logic that, given an `Expense`, automatically determine which should be the corresponding category to use when publishing to Airtable or YNAB. Similarly, it also implements logic that, given a `BankTransaction`, automatically identifies which is the corresponding Expense category.

## How
### Rules
We have custom `Rule` classes that, given the expense attributes, return the corresponding category for Airtable, YNAB or Expense. Each of this classes is supposed to represent a type of expense. See them in the `Finance::ExpenseClassification::Rules` module

Additionally, in order to make the process to add new rules as simple as possible, we have an `ALL` collection that automatically identifies the rule classes and returns a list with all of them.

### Classification
Each entity that needs a Category has a corresponding `CategoryClassifier` class (examples: `Airtable::CategoryClassifier` or `Finance::CategoryClassifier`). These classes implement a class method `.category_id_for` that, given the entity, returns the corresponding category identifier.

### Wiring
For the moment, we haven't wired these components to any part of the application. We'll handle that in a separate PR, to refactor a bit the logic and avoid concentrating too many changes in this PR.


## Commits
* Moving Airtable and YNAB classes to the Models dir

These classes are not really services, but representations of data in
Airtable and YNAB. It makes more sense for them to be in the Models
directory, instead on the Services one

* Encapsulating Airtable expense publication

Creating a new class, Airtable::ExpensePublisher, that encapsulates the
responsibility of publishing a finance expense into Airtable. This will
make it easier to reuse it across the codebase

* Encapsulating YNAB expense publication

Following the same strategy than on the previous commit, but this time
with YNAB expenses

* Adding additional expense categories

* Automating YNAB Expense classification

Adding a Ynab::CategoryClassifier class to automatically obtain the
category ID for an expense based on its amount and description. Added
rules with some of the most common expense types

* Automating BankTransaction classification

Extending the ExpenseClassification Rules to add an `expense_category`
method that returns the corresponding ExpenseCategory for that expense.

Creating the Finance::CategoryClassifier class, that given a specific
BankTransaction, analyses all the rules and returns the matching
category. To do this, we have created a `BankTransactionExpenseAdapter`
class, that allows us to access the BankTransaction attributes using the
Expense interface.

* Adding Expense Classification for Airtable

Repeating the implementation we used for YNAB and BankTransactions, but
this time with Airtable. Extending the rules to return Airtable ID and
implementing a CategoryClassifier class

* Classifying Expenses that don't match any rule

Adding explicit logic to manage expenses that don't match any direct
rule, we will use a specific category that identifies them as
unclassified.