Wiring Automated Transaction Classification (#58)
* Updating fixtures path

Moving the SNS event fixtures to the `aws/sns` folder and adding
fixtures for the DynamoDB stream events in `aws/dynamodb`

* Fixing BankTransactionExpenseAdapter

- Fix `amount` method to make the conversion to expense when required
- Adding methods to set fields, and raising an exception to make
explicit that those aren't supported

* Creating Expenses when a BankTransaction is created

Modifying BankTransactionCreatedJob so when a Bank Transaction is
created, we create an Expense from its data. To help with the debugging during the initial phases, we will be logging the created Expense in Telegram

* Creating ExpenseCreatedJob

This commit splits `CreateExpenseJob` into two different jobs:

- **MoneySpentJob**: reads MoneySpent SNS events and creates a Expense
from them
- **ExpenseCreatedJob**: given a created Expense, publishes it in
CloudWatch, Airtable and YNAB

The main goal of this change is to enable publication of expenses no
matter how they were created. This will trigger, for example, when the
expenses are created from a BankTransaction via BankTransactionCreatedJob

As we yet don't have any real DynamoDB event triggered by an Expense
creation, we have adapted the one we get from BankTransaction creation