Fixing Expense Metric publication
We added a new Expense Category (Subscription) but we didn't add a
mapping for it in `ExpenseMetricAdapter`. Therefore, the dimension was
being published with `nil` value in the dimension, generating errors
when using PutMetricData.

```
ArgumentError: missing required parameter params[:metric_data][0][:dimensions][0][:value]
```

This commit improves the category mapping so that first tries a direct
mapping, and if it's empty, it just uses the category but capitalized.
With this, even if we add new categories we won't need to specify a
mapping manually