Empty commit to trigger a CI build
The previous CI build took a long time due to dependency installation.
Triggering a new build without any changes to get a better estimate of
build times when we use dependency cache.

Additionally, the build for Ruby 2.7 failed due to permission issues
when trying to extract the cache, which shouldn't happen because we're
only using common CircleCI steps and therefore everything should have
the correct permissions. Trying to see if everything works as expected
this time