Dexcom backfilling (#59)
* Backfill Dexcom data

Modifying Health::GetLiveDexcomDataJob to retrieve the datapoints in the
last 24h, instead of just the last datapoint. This will allow us to
retrieve backfilled data: data that may not have been available
instantly (maybe because the sensor lost connectivity, or because the
transmitter didn't publish it on time) but that gets eventually
published

* Allowing batch creation of GlucoseValues

Now that we're getting datapoints for an entire day, the job execution
is much slower as it needs to do one DynamoDB call per datapoint. To
improve this, we will save all the records in a single call, using the
`.import` method instead of `.new` or `.create`

See https://github.com/Dynamoid/dynamoid#object-creation for more
information about the method