Getting only the last 3 hours of Dexcom data
As we are publishing all the available datapoints every 5 minutes, we
are getting closer to CloudWatch Free Tier limit for number of
PutMetricData requests. To prevent getting over this limit, we'll reduce
the publication time.

Dexcom transmitter only holds 3 hours of data, so any data that is still
unavailable after that will never be available. Therefore, it doesn't
make sense to try to query it.