Migrating AIB integration to Plaid (#66)
* Start preparing Plaid

Start preparing the migration to Plaid:

- Add dependency
- Create a factory to configure the client
- Create FactoryBot factories to help during testing

* Happy path for AIB with Plaid

Migrating the happy path of the AIB service class to use Plaid instead
of Truelayer. All the tests for that part of the code are passing, but
the ones related with non-happy path (e.g. exceptions, auth errors) are
failing

* Exception handling for AIB with Plaid

Adapting Finance::Aib::Service to handle non-happy cases of the Plaid
integration, like token expiration or request errors

* Adapting AIB Transaction Builder to Plaid format