Bugfix: Correctly specify token for Plaid (#67)
The previous implementation of the Plaid integration failed because we
were sending the entire AuthToken instead of just the access_token. This
commit fixes that.

Additionally, we fix a bug in the custom have_fields RSpec matcher,
which wasn't correctly failing if some of the fields weren't
matching. We have also added tests for the matcher to catch similar
errors earlier in the future