Adding Redis cache to avoid datapoint duplication
Closes #53

With each execution of Health::GetLiveDexcomDataJob, we retrieve all the
datapoints corresponding to the last 3h and publish them in CloudWatch.
This is helpful to avoid missing datapoints in case the datapoints get
published to Dexcom with a delay (e.g. the transmitter slows down or
loses connection for a few minutes). However, it also means that when
everything works correctly, we will be publishing each datapoint
multiple times, which is an inefficient use of CloudWatch

Here we resolve that problem by adding a Redis DB that we will use as a
deduplication cache: whenever we publish a glucose value in CloudWatch,
we will store its timestamp in Redis. On each execution, before getting
to the CloudWatch publication part, we will check if the timestamps are
present in Redis, and filter them out if that's the case

Also, as we don't want this to be a critical blocker, we will just
ignore Redis in case it raises errors, and assume that the datapoint is
not published