Exclude pending transactions from AIB
Currently we are processing AIB transactions twice: once when they
happen for the first time, and another one when they are confirmed. This
happens because most AIB transactions are initially marked as pending,
and only get confirmed after 1/2 days.

The goal of this change is to stop processing transactions that haven't
been confirmed yet, on the one hand to avoid this double processing that
we mentioned, and on the other because the details might not be final
and therefore the accumulated data will be incorrect at the end of the
month