Expiring Redis keys
Given that we only use Redis keys for a max of three hours, we will set
an expiration time on them to make sure that they don't keep
accumulating after they are no longer needed. This will help us save
space in our Redis cluster

To add an extra safety margin, we will expire keys after six hours
instead of three as the current job reads. This should never be needed,
but given that we are supposed to have a small number of keys, the extra
space these keys will generate doesn't really matter