Downgrading puma and upgrading byebug to use versions available in Lambdagems
puma and byebug require native extensions to work, which means that they
have to be compiled on an environment similar to where they will be run.
Jets manages this by using www.serverlessgems.com, a service that contains
gems already compiled for Lambda.

However, the last versions of puma is not yet available in
Serverlessgems, and therefore the deploy fails. We'll pin them to the
previous version that we were using until they are available.

Additionally, we will remove the constraint in byebug. The latest
version was not available in the past (see
https://github.com/AlexGascon/alexgascon-api/commit/5cbdbe54821f716e6fa9f5d55d088884faf32744),
but is available now