Alerting when Plaid consent expires
Plaid uses OpenBanking APIs, so user consent needs to be
renewed periodically for security reasons (in Europe at least). The
problem is that Plaid does not emit any kind of error when the consent
expires, it just stops retrieving transactions.

The goal of this change is to check the consent expiration each time that
we get a response from Plaid, and confirm that it's still valid. If not,
raise an Auth exception, so our monitoring system detects the problem