Getting 24h of Dexcom values instead of 3h
Previously, we were getting 3h because that is the max amount of data that is stored in the transmitter locally, if it's not able to send it to the cloud. That means that if there was data that wasn't published in the previous 3h, that data was going to be lost.

However, now the Dexcom Share app seems to be misbehaving, and even if the data is successfully published from the transmitter to the Dexcom app, it's not available on Share on time. It eventually is available, but that might happen out of the 3h query period that we were using, and therefore we won't get the data.

The goal of this change is to change the query period to 24h, the max allowed by Dexcom, to ensure that we get all the data that is eventually published, even if it's with a lot of delay. This will give us at least allow to have it for historic purposes (e.g. check the specific value of glucose a week ago)