Feature/classification rules improvements (#86)
* Adding new categories + Minor changes in existing ones

Adding categories for other recurrent expenses

Modifying the amount in some categories (Spotify, Pepephone) as the
amount of the recurrent expense has changed. In Pepephone we remove the
amount completely because the exact value can change depending on
variables like the number of calls, SMSs, etc

* Allowing more complex expense classification rules

Modifying base_rule.rb to enable the possibility of overriding
amount_matches? and description_matches? directly. This will allow us to
create more complex matching rules, that work with the expense directly
and allow everything that can be done with a method instead of limiting
us to an equality comparison

Added a rule_implementation_error? check to ensure that at least one of
the matching characteristics is overriden (otherwise, the rule will
match everything).

Added a Supermercado rule that uses this new structure, and modifying
the Deliveroo rule to do the same

_Note: there was nothing forbidding us from overriding amount_matches?
and description_matches? before, but now with the implementation error
check we have a guardrail that makes it explicit and protects us from mistakes_

* Recording data of failed RSpec tests

This allows us to use the --only-failures flag, repeating only the
failed tests from the last execution